+++ EVTCAnalytics/GameData/GameBuilds.cs
@@ -9,5 +9,10 @@ namespace GW2Scratch.EVTCAnalytics.GameData
 		/// The release of The Key of Ahdashim, which included a significant rework to raid rewards.
 		/// </summary>
 		public static int AhdashimRelease = 97235;
+
+		/// <summary>
+		/// The release of Cosmic Observatory Challenge Mode, which revamped the health of the normal mode as well.
+		/// </summary>
+		public static int CosmicObservatoryCMRelease = 153978;
 	}
 }